Related to #2 Now it's time to write those tests and the implementation.
Write the test before writing the implementation code .

The benefits of doing this are as follows—it ensures that testable code is written and ensures that every line of code gets tests written for it.

By writing or modifying the test first, the developer is focused on requirements before starting to work on a code. This is the main difference when compared to writing tests after the implementation is done. An additional benefit is that with tests first, we are avoiding the danger that the tests work as quality checking instead of quality assurance.

-------------

One more time, notes how TDD forces you to think from the requirements, and on design
Right now, the code even compiles!